Report, Assign and Unassign now subclasses of EasymanageCurationWorkApp
